test/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
requests==2.31.0
python-dotenv==1.0.1
numpy==1.26.4
scikit-learn==1.4.0
